requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pdfplumber==0.11.0
supabase==2.23.2
python-dotenv==1.0.1
